Move `lookupInput` and `lookupFile` from `servant-multipart` to `servant-multipart-api`
This adds one extra dependency to `servant-multipart-api`, namely `string-conversions`, but this library is small and it's dependencies are already dependencies of `servant-multipart-api` anyway.

It means that downstream libraries that call these functions don't have to transitively depend on `servant-server`, which itself has many dependencies, a number of which are not able to be built under WASM (such as `network`).
