+++ servant-multipart-api/servant-multipart-api.cabal
@@ -40,6 +40,7 @@ library
   -- other dependencies
   build-depends:
       servant             >=0.16     && <0.21
+    , string-conversions  >=0.4.0.1  && <0.5
 
   -- servant-0.19 dropped support for GHC-8.4 (latest GHCJS version),
   -- due to QuantifiedConstraints
